add AdaBoost wrapper

+++ include/distance_weighting/distance_weighting.hpp
@@ -15,27 +15,27 @@ namespace alfarobi_v8{
 
 class DistanceWeightingSearch{
 public:
-    using max_val_t = std::vector<double>;
+    using max_val_t = std::vector<float>;
     using max_point_t = std::vector<cv::Point>;
 
 public:
     auto distanceWeighting(const cv::Mat &in){
-        cv::Mat weighting(cv::Mat::zeros(in.size(), CV_64FC1));
+        cv::Mat weighting(cv::Mat::zeros(in.size(), CV_32FC1));
         auto num_rows(in.rows);
         auto num_cols(in.cols);
         for(auto i(1); i < (num_rows - 1); i++){		
             const uchar *in_ptr1 = in.ptr<uchar>(i-1);
             const uchar *in_ptr2 = in.ptr<uchar>(i);
             const uchar *in_ptr3 = in.ptr<uchar>(i+1);
-            double *weighting_ptr1 = weighting.ptr<double>(i-1);
-            double *weighting_ptr2 = weighting.ptr<double>(i);
-            double *weighting_ptr3 = weighting.ptr<double>(i+1);
+            float *weighting_ptr1 = weighting.ptr<float>(i-1);
+            float *weighting_ptr2 = weighting.ptr<float>(i);
+            float *weighting_ptr3 = weighting.ptr<float>(i+1);
             for(auto j(1); j < (num_cols - 1); j++){
                 if(in_ptr2[j]){
                     if(in_ptr3[j-1] !=0 && in_ptr3[j] !=0 && in_ptr3[j+1] !=0 && 
                         in_ptr2[j-1] !=0 && in_ptr2[j] !=0 && in_ptr2[j+1] !=0 &&
                         in_ptr1[j-1] !=0 && in_ptr1[j] !=0 && in_ptr1[j+1] !=0){					
-                        double min_dist = fminf(fminf(weighting_ptr2[j-1], weighting_ptr1[j-1]),
+                        float min_dist = fminf(fminf(weighting_ptr2[j-1], weighting_ptr1[j-1]),
                                                 fminf(weighting_ptr1[j], weighting_ptr1[j+1]));				
                         weighting_ptr2[j] = min_dist + 1;
                     }else{
@@ -74,7 +74,7 @@ public:
                 //-- remove visited candidate
                 cv::circle(dt, ctr, maxv, cv::Scalar(0), cv::FILLED);
                 max_point.push_back(ctr);
-                max_val.push_back(maxv);                
+                max_val.push_back(static_cast<float>(maxv));                
             }else {
                 for(auto j(0); j < MAX_CANDIDATE; j++){
                     cv::Mat interest(dt, r);
@@ -86,7 +86,7 @@ public:
                     //-- remove visited candidate               
                     cv::circle(dt, ctr, maxv, cv::Scalar(0), cv::FILLED);
                     max_point.push_back(ctr);        		
-                    max_val.push_back(maxv);
+                    max_val.push_back(static_cast<float>(maxv));
                 }
             }
         }
